+++ engine/bonus-malus/castling.hpp
@@ -8,11 +8,11 @@ namespace engine {
 // CASTLING
 // ===================================================
 
-inline static constexpr int64_t CASTLING_BONUS = 30;
+inline static constexpr int32_t CASTLING_BONUS = 30;
 
-inline static constexpr int64_t KING_NON_CASTLING_PENALTY = 10;
-inline static constexpr int64_t KING_LOST_CASTLING_RIGHTS_PENALTY = 25;
-inline static constexpr int64_t LOSS_OF_CASTLING_PENALTY = 35;
+inline static constexpr int32_t KING_NON_CASTLING_PENALTY = 10;
+inline static constexpr int32_t KING_LOST_CASTLING_RIGHTS_PENALTY = 25;
+inline static constexpr int32_t LOSS_OF_CASTLING_PENALTY = 35;
 
 } // namespace engine
