+++ engine/bonus-malus/rook_endgame.hpp
@@ -9,10 +9,10 @@ namespace engine {
 // ===================================================
 
 // bonus when opponent king is near edge
-inline static constexpr int64_t ROOK_EG_EDGE_BONUS = 35;
+inline static constexpr int32_t ROOK_EG_EDGE_BONUS = 35;
 
 // bonus for active coordination
-inline static constexpr int64_t ROOK_EG_PRESSURE_BONUS = 20;
+inline static constexpr int32_t ROOK_EG_PRESSURE_BONUS = 20;
 
 } // namespace engine
